+++ src/HorseCollisionMod.lua
@@ -66,11 +66,11 @@
 --
 -- @module HorseCollisionMod
 -- @author jrandall54
--- @release 4.2.5
+-- @release 4.2.6
 
 HorseCollisionMod = {}
 
-HorseCollisionMod.Version = "4.2.5"
+HorseCollisionMod.Version = "4.2.6"
 
 --- Loop generation counter, deliberately kept outside the table above.
 --